+++ CHANGES.md
@@ -1,3 +1,19 @@
+1.1.0 (2022-04-19)
+------------------
+- Added a new hip2 parser, in special_parse.py, that will add the residual offset and cosmic dispersion of 
+  Brandt et al. 2022 and output new IAD with new best-fit parameters. Works for 5, 7, and 9 parameter sources.
+- Added a new hip2 parser that autodetects whether the data file is the java tool or DVD. Technically, it is 
+  a parser factory.
+- Added the option for htof.main.Astrometry to load in and use the parallax factors supplied with
+the hipparcos IAD.
+- htof.special_parse.Hipparcos2Recalibrated can now write out a data file in the same format 
+  (although they do not match byte-by-byte exactly) as the hip2 java tool IAD.
+- fitting with return_all=True now returns 4 arrays, the last is the residuals (in ra and dec) of the fit.
+- New examples jupyter notebook for recalibrating the hip2 java tool IAD.
+- Astrometry() now has an along_scan_error_scaling attribute that will scale the
+  along scan errors by that factor. This is particularly useful for gaia forecasting. Setting
+  along_scan_error_scaling = 0.1 for instance, will artifically make all the Gaia GOST 
+  scans have an error of 0.1 mas (normally, these along scan errors are not provided by GOST)
 
 1.0.3 (2022-01-26)
 ------------------